+++ lib/chunk/SimpleChunkManager.cpp
@@ -16,7 +16,7 @@ void SimpleChunkManager::setChunk(int_fast64_t chunkX, int_fast64_t chunkZ, Chun
   chunks.insert({morton2d_encode(chunkX, chunkZ), chunk});
 }
 
-MinecraftBlock SimpleChunkManager::getBlockAt(int_fast64_t x, int_fast16_t y, int_fast64_t z) {
+MinecraftBlock SimpleChunkManager::getBlockAt(int_fast64_t x, int_fast32_t y, int_fast64_t z) {
   Chunk *chunk;
   if (isInWorld(x, y, z) && (chunk = getChunk(x >> 4, z >> 4)) != nullptr) {
     return MinecraftBlock(chunk->getFullBlock(static_cast<int_fast8_t>(x & 0xf), y, static_cast<int_fast8_t>(z & 0xf)));
@@ -25,7 +25,7 @@ MinecraftBlock SimpleChunkManager::getBlockAt(int_fast64_t x, int_fast16_t y, in
   return MinecraftBlock((Block) 0);
 }
 
-void SimpleChunkManager::setBlockAt(int_fast64_t x, int_fast16_t y, int_fast64_t z, MinecraftBlock block) {
+void SimpleChunkManager::setBlockAt(int_fast64_t x, int_fast32_t y, int_fast64_t z, MinecraftBlock block) {
   Chunk *chunk;
 
   if ((chunk = getChunk(x >> 4, z >> 4)) != nullptr) {
@@ -37,15 +37,15 @@ void SimpleChunkManager::setBlockAt(int_fast64_t x, int_fast16_t y, int_fast64_t
   }
 }
 
-bool SimpleChunkManager::isInWorld(int_fast64_t x, int_fast16_t y, int_fast64_t z) const {
+bool SimpleChunkManager::isInWorld(int_fast64_t x, int_fast32_t y, int_fast64_t z) const {
   return x <= INT32_MAX && x >= INT32_MIN && y < maxY && y >= minY && z <= INT32_MAX && z >= INT32_MIN;
 }
 
-int_fast16_t SimpleChunkManager::getMinY() const {
+int_fast32_t SimpleChunkManager::getMinY() const {
   return minY;
 }
 
-int_fast16_t SimpleChunkManager::getMaxY() const {
+int_fast32_t SimpleChunkManager::getMaxY() const {
   return maxY;
 }
