+++ lib/chunk/SimpleChunkManager.h
@@ -8,18 +8,18 @@
 
 class SimpleChunkManager {
  public:
-  SimpleChunkManager(const int_fast16_t minHeight, const int_fast16_t maxHeight) : minY(minHeight), maxY(maxHeight) {}
+  SimpleChunkManager(const int_fast32_t minHeight, const int_fast32_t maxHeight) : minY(minHeight), maxY(maxHeight) {}
 
   auto setChunk(int64_t chunkX, int64_t chunkZ, Chunk *chunk) -> void;
   auto getChunk(int64_t chunkX, int64_t chunkZ) -> Chunk *;
 
-  auto getBlockAt(int_fast64_t x, int_fast16_t y, int_fast64_t z) -> MinecraftBlock;
-  auto setBlockAt(int_fast64_t x, int_fast16_t y, int_fast64_t z, MinecraftBlock block) -> void;
+  auto getBlockAt(int_fast64_t x, int_fast32_t y, int_fast64_t z) -> MinecraftBlock;
+  auto setBlockAt(int_fast64_t x, int_fast32_t y, int_fast64_t z, MinecraftBlock block) -> void;
 
-  [[nodiscard]]  auto isInWorld(int_fast64_t x, int_fast16_t y, int_fast64_t z) const -> bool;
+  [[nodiscard]]  auto isInWorld(int_fast64_t x, int_fast32_t y, int_fast64_t z) const -> bool;
 
-  [[nodiscard]] auto getMinY() const -> int_fast16_t;
-  [[nodiscard]] auto getMaxY() const -> int_fast16_t;
+  [[nodiscard]] auto getMinY() const -> int_fast32_t;
+  [[nodiscard]] auto getMaxY() const -> int_fast32_t;
 
   [[nodiscard]] auto getChunks() const -> std::map<uint_fast64_t, Chunk *>;
 
@@ -28,8 +28,8 @@ class SimpleChunkManager {
  private:
   std::map<uint_fast64_t, Chunk *> chunks;
 
-  int_fast16_t minY;
-  int_fast16_t maxY;
+  int_fast32_t minY;
+  int_fast32_t maxY;
 };
 
 #endif // EXT_NOISELIB_LIB_CHUNK_SIMPLECHUNKMANAGER_H_
